// admin/src/app/models/tour.model.ts
export class Tour {
    Id: null;
    LocationID: number;
    CountryID: number;
    CountryLocationID: number;
    TourPackageID: number;
    TourPackageLocationID: number;
    TourProgramID: number;
    TourProgramLocationID: number;
    CityID: number;
    CityLocationID: number;
    TourTitle: string;
    TourLocation: String;
    TourDescription: string;
    SeasonStartDate: string;
    SeasonEndDate: string;
    DurationInDays: number;
    DurationInNights: number;
    FrontPage:boolean;
    FlightIncluded:boolean;
    LanguageValue: string;
    CreatedOn: null;
};